requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
monai>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Medical Imaging
pydicom>=2.4.0
nibabel>=5.1.0
SimpleITK>=2.3.0

# Training & Monitoring
tensorboard>=2.15.0
mlflow>=2.9.0
tqdm>=4.66.0

# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-multipart>=0.0.6
aiofiles>=23.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
Pillow>=10.0.0
matplotlib>=3.8.0
pandas>=2.1.0
h5py>=3.10.0
onnx>=1.15.0
onnxruntime-gpu>=1.16.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.23.0

# MLOps
dvc>=3.30.0
